+++ src/OrdersList/hooks/useBuildQuery/useBuildQuery.test.js
@@ -1,9 +1,12 @@
 import queryString from 'query-string';
 
 import { renderHook } from '@folio/jest-config-stripes/testing-library/react';
-import { ORDER_STATUSES } from '@folio/stripes-acq-components';
+import {
+  CUSTOM_FIELDS_FILTER,
+  CUSTOM_FIELDS_FIXTURE,
+  ORDER_STATUSES,
+} from '@folio/stripes-acq-components';
 
-import { CUSTOM_FIELDS } from 'fixtures/customFields';
 import { FILTERS } from '../../constants';
 import { useBuildQuery } from './useBuildQuery';
 
@@ -19,16 +22,16 @@ describe('useBuildQuery', () => {
   });
 
   it('returned function should return proper query for custom field type MULTI_SELECT_DROPDOWN', () => {
-    const { result } = renderHook(() => useBuildQuery(CUSTOM_FIELDS));
-    const query = result.current(queryString.parse(`?${FILTERS.CUSTOM_FIELDS}.multiselect=opt_0`));
+    const { result } = renderHook(() => useBuildQuery(CUSTOM_FIELDS_FIXTURE));
+    const query = result.current(queryString.parse(`?${CUSTOM_FIELDS_FILTER}.multiselect=opt_0`));
 
-    expect(query).toBe(`(${FILTERS.CUSTOM_FIELDS}.multiselect=="*opt_0*") ${SORT_STRING}`);
+    expect(query).toBe(`(${CUSTOM_FIELDS_FILTER}.multiselect=="*opt_0*") ${SORT_STRING}`);
   });
 
   it('returned function should return proper query for custom field type SINGLE_SELECT_DROPDOWN', () => {
-    const { result } = renderHook(() => useBuildQuery(CUSTOM_FIELDS));
-    const query = result.current(queryString.parse(`?${FILTERS.CUSTOM_FIELDS}.singleselect=opt_0`));
+    const { result } = renderHook(() => useBuildQuery(CUSTOM_FIELDS_FIXTURE));
+    const query = result.current(queryString.parse(`?${CUSTOM_FIELDS_FILTER}.singleselect=opt_0`));
 
-    expect(query).toBe(`(${FILTERS.CUSTOM_FIELDS}.singleselect=="opt_0") ${SORT_STRING}`);
+    expect(query).toBe(`(${CUSTOM_FIELDS_FILTER}.singleselect=="opt_0") ${SORT_STRING}`);
   });
 });