+++ src/components/POLine/POLineView.js
@@ -35,6 +35,7 @@ import {
 } from '@folio/stripes/smart-components';
 
 import {
+  CUSTOM_FIELDS_ORDERS_BACKEND_NAME,
   DonorsListContainer,
   FundDistributionView,
   handleKeyCommand,
@@ -61,7 +62,6 @@ import {
   ORDERS_ROUTE,
   REEXPORT_SOURCES,
   ORDER_LINES_ROUTE,
-  CUSTOM_FIELDS_BACKEND_MODULE_NAME,
 } from '../../common/constants';
 import { useExportHistory } from '../../common/hooks';
 import { isOngoing } from '../../common/POFields';
@@ -531,7 +531,7 @@ const POLineView = ({
 
                 <ViewCustomFieldsRecord
                   accordionId="customFieldsPOLine"
-                  backendModuleName={CUSTOM_FIELDS_BACKEND_MODULE_NAME}
+                  backendModuleName={CUSTOM_FIELDS_ORDERS_BACKEND_NAME}
                   customFieldsValues={customFieldsValues}
                   entityType="po_line"
                 />